feat: save checkpoints atomically in API trigger_ingest
Move build_checkpoint to adapters::repo as a shared public function
(was duplicated in CLI). API trigger_ingest now calls
save_transactions_and_checkpoint when a checkpoint can be built,
matching the CLI's atomic save behavior.
